// Nexo/src/types.ts
export type SubscriptionCategory =
  | 'entertainment'
  | 'productivity'
  | 'health'
  | 'utilities'
  | 'cloud'
  | 'other'

export type SubscriptionStatus = 'active' | 'cancelled'

export interface Subscription {
  id: string
  name: string
  logo: string
  category: SubscriptionCategory
  amount: number
  frequency: 'monthly' | 'annual'
  nextBillingDate: string
  status: SubscriptionStatus
  merchantName: string
}